requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
langchain
langchain-openai
chromadb